+++ src/main/java/com/fourthread/ozang/module/domain/security/user/UserDetailsImpl.java
@@ -15,7 +15,7 @@ public class UserDetailsImpl implements UserDetails {
 
   @Override
   public Collection<? extends GrantedAuthority> getAuthorities() {
-    return List.of(new SimpleGrantedAuthority(user.getRole().name()));
+    return List.of(new SimpleGrantedAuthority("ROLE_" + user.getRole().name()));
   }
 
   @Override
